mcp_servers/social_mcp/requirements.txt
praw 
pydantic 
fastapi 
python-dotenv
Tweepy
TikTokApi